Migrate Mutation.updateProjectRegistryModel
